refactor(test): split codec conformance per module
One 271-line file held two unrelated suites. `hex.ts` and `base64.ts` are
separate modules and their tier suites share nothing but the input generators,
which move to the harness where the next codec can reach them.

Both projects now collect `*.conformance.ts` directly, so a suite is written
once and each runtime runs whichever tiers it has. That removes the two shim
files whose only job was to match a project's filename glob.
