--- src/core/internal/codec/_test/conformance.browser.test.ts
@@ -1,3 +1,0 @@
-// Browser: exercises the JS loop and the native `Uint8Array` methods. No
-// `Buffer` here, so that tier is covered by the Node run of the same suite.
-import './conformance.js'
