--- src/core/internal/codec/_test/conformance.test.ts
@@ -1,4 +1,0 @@
-// Node: exercises the JS loop and, for hex, `Buffer`. Node has neither
-// `Uint8Array.fromHex` nor `fromBase64`, so the native tiers are covered by the
-// browser run of the same suite.
-import './conformance.js'
